[+] Sanity check that the log actually exists

+++ .github/scripts/x86_64_Linux/gen_meta.sh
@@ -121,18 +121,20 @@ echo -e "\n\n [+] Started Metadata Update at :: $(TZ='Asia/Kathmandu' date +'%A,
             #Web URLs
              jq --arg BIN "$BIN" --arg WEB_URL "$WEB_URL" '.[] |= if .name == $BIN then . + {web_url: $WEB_URL} else . end' "$TMPDIR/METADATA.json" > "$TMPDIR/METADATA.tmp" && mv "$TMPDIR/METADATA.tmp" "$TMPDIR/METADATA.json"
             #Build Log
-             LOG_MATCH="$(echo "${BUILD_SCRIPT}" | sed 's|https://github.com/Azathothas/Toolpacks/tree/main|https://pub.ajam.dev/repos/Azathothas/Toolpacks|')" && export LOG_MATCH="${LOG_MATCH}"
-             #LOG_BIN="$(echo "${BIN}" | sed 's/\.no_strip$//')" && export LOG_BIN="${LOG_BIN}"
-             LOG_BIN="$(echo "${BIN}")" && export LOG_BIN="${LOG_BIN}"
-             LOG_BEGIN="$(grep -nE "^\[\+\] (Building|Fetching) \: ${LOG_MATCH}" "$SYSTMP/BUILD.log" | head -n 1 | cut -d: -f1)" && export LOG_BEGIN="${LOG_BEGIN}"
-             LOG_END="$(awk -v start="$LOG_BEGIN" 'NR > start && /.*Completed \(Building\|Fetching\).*/ {print NR; exit}' "$SYSTMP/BUILD.log")" && export LOG_END="${LOG_END}"
-             if [ -n "${LOG_BEGIN}" ] && [ -n "${LOG_END}" ]; then
-               sed -n "${LOG_BEGIN},${LOG_END}p" "$SYSTMP/BUILD.log" > "$SYSTMP/BIN_LOGS/${BIN}.log.txt"
-               cat "$SYSTMP/BIN_LOGS/${BIN}.log.txt" > "$SYSTMP/BIN_LOGS/$(echo ${BUILD_URL} | sed -n 's|.*\/bins/\(.*\)\.yaml$|\1|p').log.txt"
-               BUILD_LOG="$(jq --arg BIN "$BIN" -r '.[] | select(.name == $BIN) | .download_url' "$TMPDIR/METADATA.json" | sed 's/\.no_strip$//').log.txt"
-               jq --arg BIN "$BIN" --arg BUILD_LOG "$BUILD_LOG" '.[] |= if .name == $BIN then . + {build_log: $BUILD_LOG} else . end' "$TMPDIR/METADATA.json" > "$TMPDIR/METADATA.tmp" && mv "$TMPDIR/METADATA.tmp" "$TMPDIR/METADATA.json"
+             if [ -f "$SYSTMP/BUILD.log" ] && [ $(stat -c%s "$SYSTMP/BUILD.log") -gt 10240 ]; then
+               LOG_MATCH="$(echo "${BUILD_SCRIPT}" | sed 's|https://github.com/Azathothas/Toolpacks/tree/main|https://pub.ajam.dev/repos/Azathothas/Toolpacks|')" && export LOG_MATCH="${LOG_MATCH}"
+               #LOG_BIN="$(echo "${BIN}" | sed 's/\.no_strip$//')" && export LOG_BIN="${LOG_BIN}"
+               LOG_BIN="$(echo "${BIN}")" && export LOG_BIN="${LOG_BIN}"
+               LOG_BEGIN="$(grep -nE "^\[\+\] (Building|Fetching) \: ${LOG_MATCH}" "$SYSTMP/BUILD.log" | head -n 1 | cut -d: -f1)" && export LOG_BEGIN="${LOG_BEGIN}"
+               LOG_END="$(awk -v start="$LOG_BEGIN" 'NR > start && /.*Completed \(Building\|Fetching\).*/ {print NR; exit}' "$SYSTMP/BUILD.log")" && export LOG_END="${LOG_END}"
+               if [ -n "${LOG_BEGIN}" ] && [ -n "${LOG_END}" ]; then
+                 sed -n "${LOG_BEGIN},${LOG_END}p" "$SYSTMP/BUILD.log" > "$SYSTMP/BIN_LOGS/${BIN}.log.txt"
+                 cat "$SYSTMP/BIN_LOGS/${BIN}.log.txt" > "$SYSTMP/BIN_LOGS/$(echo ${BUILD_URL} | sed -n 's|.*\/bins/\(.*\)\.yaml$|\1|p').log.txt"
+                 BUILD_LOG="$(jq --arg BIN "$BIN" -r '.[] | select(.name == $BIN) | .download_url' "$TMPDIR/METADATA.json" | sed 's/\.no_strip$//').log.txt"
+                 jq --arg BIN "$BIN" --arg BUILD_LOG "$BUILD_LOG" '.[] |= if .name == $BIN then . + {build_log: $BUILD_LOG} else . end' "$TMPDIR/METADATA.json" > "$TMPDIR/METADATA.tmp" && mv "$TMPDIR/METADATA.tmp" "$TMPDIR/METADATA.json"
+               fi
              else
-               jq --arg BIN "$BIN" --arg BUILD_LOG "https://bin.ajam.dev/x86_64_Linux/BUILD.log.txt" '.[] |= if .name == $BIN then . + {build_log: $BUILD_LOG} else . end' "$TMPDIR/METADATA.json" > "$TMPDIR/METADATA.tmp" && mv "$TMPDIR/METADATA.tmp" "$TMPDIR/METADATA.json"
+                 jq --arg BIN "$BIN" --arg BUILD_LOG "https://bin.ajam.dev/x86_64_Linux/BUILD.log.txt" '.[] |= if .name == $BIN then . + {build_log: $BUILD_LOG} else . end' "$TMPDIR/METADATA.json" > "$TMPDIR/METADATA.tmp" && mv "$TMPDIR/METADATA.tmp" "$TMPDIR/METADATA.json"
              fi
             #Build_Script
              jq --arg BIN "$BIN" --arg BUILD_SCRIPT "$BUILD_SCRIPT" '.[] |= if .name == $BIN then . + {build_script: $BUILD_SCRIPT} else . end' "$TMPDIR/METADATA.json" > "$TMPDIR/METADATA.tmp" && mv "$TMPDIR/METADATA.tmp" "$TMPDIR/METADATA.json"
